pyproject.toml
[tool.poetry]
name = "colocanalyse"
version = "0.1.0"
description = ""
authors = ["Abbie H Davidson"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.8"
readlif = "0.6.5"
numpy = "2.2.0"
trackpy = "0.6.4"
pandas = "2.2.3"
tqdm = "4.67.1"

hydra-core = "1.3.2"
scikit-image = "0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
